pyproject.toml
[tool.poetry]
name = "trend-based-content-generator-crew"
version = "0.1.0"
description = "Generate articles that capture current trends within products"
authors = ["Srinivas Aripirala <srinivas.aripirala@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.11.9,<3.12"
crewai = "^0.30.11"
crewai-tools = "0.2.6"
python-dotenv = "1.0.1"
opencv-python = "4.8.1.78"
google-search-results = "2.4.2"
unstructured = "0.14.6"
tools = "^0.1.9"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"